moved current_tx out of fields/
