+++ xrpl-common-stdlib/src/ctx/mod.rs
@@ -3,7 +3,7 @@
 //! This module exposes the [`SmartFeatureContext`] trait — the narrow context
 //! type implemented by feature-specific contexts (e.g., `EscrowFinishContext`)
 
-use crate::fields::current_tx::traits::TransactionCommonFields;
+use crate::current_tx::traits::TransactionCommonFields;
 
 /// Narrow context trait shared by all Smart Escrow and Smart Contract entry points.
 ///
