+++ xrpl-common-stdlib/src/current_tx/traits.rs
@@ -38,7 +38,7 @@
 //! - **PublicKey**: 33-byte compressed public keys
 //! - **TransactionType**: Enumerated transaction type identifiers
 
-use crate::fields::current_tx::{get_field, get_field_optional};
+use crate::current_tx::{get_field, get_field_optional};
 use crate::host::Result;
 use crate::sfield;
 use crate::types::account_id::AccountID;
@@ -301,7 +301,7 @@ pub trait TransactionCommonFields {
 
 #[cfg(test)]
 mod tests {
-    use crate::fields::current_tx::traits::TransactionCommonFields;
+    use crate::current_tx::traits::TransactionCommonFields;
     use crate::host::host_bindings_trait::MockHostBindings;
     use crate::sfield::SField;
     use mockall::predicate::{always, eq};
@@ -329,9 +329,9 @@ mod tests {
     mod transaction_common_fields {
 
         mod optional_fields {
-            use crate::fields::current_tx::traits::TransactionCommonFields;
-            use crate::fields::current_tx::traits::tests::TestTransaction;
-            use crate::fields::current_tx::traits::tests::expect_tx_field;
+            use crate::current_tx::traits::TransactionCommonFields;
+            use crate::current_tx::traits::tests::TestTransaction;
+            use crate::current_tx::traits::tests::expect_tx_field;
             use crate::host::error_codes::{FIELD_NOT_FOUND, INTERNAL_ERROR, INVALID_FIELD};
             use crate::host::host_bindings_trait::MockHostBindings;
             use crate::host::setup_mock;
@@ -615,9 +615,9 @@ mod tests {
         }
 
         mod required_fields {
-            use crate::fields::current_tx::traits::TransactionCommonFields;
-            use crate::fields::current_tx::traits::tests::TestTransaction;
-            use crate::fields::current_tx::traits::tests::expect_tx_field;
+            use crate::current_tx::traits::TransactionCommonFields;
+            use crate::current_tx::traits::tests::TestTransaction;
+            use crate::current_tx::traits::tests::expect_tx_field;
             use crate::host::error_codes::{FIELD_NOT_FOUND, INTERNAL_ERROR, INVALID_FIELD};
             use crate::host::host_bindings_trait::MockHostBindings;
             use crate::host::setup_mock;
