+++ xrpl-common-stdlib/src/types/public_key.rs
@@ -1,4 +1,4 @@
-use crate::fields::current_tx::CurrentTxFieldGetter;
+use crate::current_tx::CurrentTxFieldGetter;
 use crate::host::field_helpers::{
     get_fixed_size_field_with_expected_bytes, get_fixed_size_field_with_expected_bytes_optional,
 };
